pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "studio_data_tools"
version = "0.1.0"
description = "Tools for generating synthetic training data for computer vision"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Studio Data Tools Team" }
]
dependencies = [
    "google-genai>=1.16.0",
    "python-dotenv>=1.0.0",
    "Pillow>=9.0.0",
    "numpy==1.26.4",
    "imgaug>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
studio_data_tools = "studio_data_tools.cli:main"

[tool.setuptools]
packages = ["studio_data_tools", "studio_data_tools.core"] 